Revert "pmdate/pmdate.c: avoid possible overflow"
This reverts commit df5aa907d08a0b0a9bb3c42fcc11c750a7f2e6d7.

Original code was correct.

+++ src/pmdate/pmdate.c
@@ -73,10 +73,10 @@ main(int argc, char *argv[])
 		now += sgn * val;
 		break;
 	    case 'M':
-		now += sgn * val * 60L;
+		now += sgn * val * 60;
 		break;
 	    case 'H':
-		now += sgn * val * 60 * 60L;
+		now += sgn * (long)val * 60 * 60;
 		break;
 	    case 'd':
 		now += sgn * val * 24 * 60 * 60;
